Added ignore_migration, miss unit test

+++ anyblok/migration.py
@@ -63,6 +63,10 @@ class MigrationReport:
         report = MigrationReport(migrationinstance, change_detected)
 
     """
+
+    def ignore_migration_for(self, table, default=None):
+        return self.migration.ignore_migration_for.get(table, default)
+
     def raise_if_withoutautomigration(self):
         if self.migration.withoutautomigration:
             raise MigrationException("The metadata and the base structue are "
@@ -92,6 +96,9 @@ class MigrationReport:
     def init_add_column(self, diff):
         self.raise_if_withoutautomigration()
         _, _, table, column = diff
+        if self.ignore_migration_for(table) is True:
+            return True
+
         self.log_names.append('Add %s.%s' % (table, column.name))
 
     def can_remove_constraints(self, name):
@@ -132,6 +139,9 @@ class MigrationReport:
 
     def init_remove_constraint(self, diff):
         _, constraint = diff
+        if self.ignore_migration_for(constraint.table) is True:
+            return True
+
         self.log_names.append('Drop constraint %s on %s' % (
             constraint.name, constraint.table))
 
@@ -155,6 +165,9 @@ class MigrationReport:
     def init_add_index(self, diff):
         self.raise_if_withoutautomigration()
         _, constraint = diff
+        if self.ignore_migration_for(constraint.table.name) is True:
+            return True
+
         columns = [x.name for x in constraint.columns]
         if self.table_is_added(constraint.table):
             return True
@@ -164,6 +177,9 @@ class MigrationReport:
 
     def init_remove_index(self, diff):
         _, index = diff
+        if self.ignore_migration_for(index.table) is True:
+            return True
+
         self.log_names.append('Drop index %s on %s' % (index.name,
                                                        index.table))
         if self.can_remove_index(index.name):
@@ -174,9 +190,15 @@ class MigrationReport:
     def init_add_fk(self, diff):
         self.raise_if_withoutautomigration()
         _, fk = diff
+        if self.ignore_migration_for(fk.table.name) is True:
+            return True
+
         from_ = []
         to_ = []
         for column in fk.columns:
+            if column.name in self.ignore_migration_for(fk.table, []):
+                return True
+
             for fk_ in column.foreign_keys:
                 from_.append('%s.%s' % (fk.table.name, column.name))
                 to_.append(fk_.target_fullname)
@@ -186,6 +208,9 @@ class MigrationReport:
 
     def init_remove_fk(self, diff):
         _, fk = diff
+        if self.ignore_migration_for(fk.table.name) is True:
+            return True
+
         for column in fk.columns:
             for fk_ in column.foreign_keys:
                 self.log_names.append('Drop Foreign keys on %s.%s => %s' % (
@@ -199,6 +224,9 @@ class MigrationReport:
     def init_add_ck(self, diff):
         self.raise_if_withoutautomigration()
         _, table, ck = diff
+        if self.ignore_migration_for(ck.table) is True:
+            return True
+
         if ck.table.schema:
             table = ck.table.schema + '.' + table
 
@@ -207,6 +235,9 @@ class MigrationReport:
 
     def init_remove_ck(self, diff):
         _, table, ck = diff
+        if self.ignore_migration_for(table) is True:
+            return True
+
         if ck['schema']:
             table = ck['schema'] + '.' + table
 
@@ -301,6 +332,12 @@ class MigrationReport:
             return True
 
     def init_modify_type(self, diff):
+        if self.ignore_migration_for(diff[2]) is True:
+            return True
+
+        if diff[3] in self.ignore_migration_for(diff[2], []):
+            return True
+
         if sgdb_in(self.migration.conn.engine, ['MySQL', 'MariaDB']):
             if isinstance(diff[5], TINYINT) and isinstance(diff[6], Boolean):
                 # Boolean are TINYINT in MySQL DataBase
@@ -316,12 +353,24 @@ class MigrationReport:
         return False
 
     def init_modify_nullable(self, diff):
+        if self.ignore_migration_for(diff[2]) is True:
+            return True
+
+        if diff[3] in self.ignore_migration_for(diff[2], []):
+            return True
+
         table = "%s.%s" % diff[1:3] if diff[1] else diff[2]
         self.log_names.append("Modify column nullable %s.%s : %s => %s" % (
             table, diff[3], diff[5], diff[6]))
         return False
 
     def init_modify_server_default(self, diff):
+        if self.ignore_migration_for(diff[2]) is True:
+            return True
+
+        if diff[3] in self.ignore_migration_for(diff[2], []):
+            return True
+
         table = "%s.%s" % diff[1:3] if diff[1] else diff[2]
         self.log_names.append("Modify column default %s.%s : %s => %s" % (
             table, diff[3], diff[5], diff[6]))
@@ -583,6 +632,7 @@ class MigrationConstraintForeignKey:
             source_schema=self.table.schema,
             referent_schema=remote_columns[0].table.schema,
             **kwargs)
+
         return self
 
     def drop(self):
@@ -751,6 +801,7 @@ class MigrationColumn:
             name = kwargs['name']
 
         if vals:
+            print(self.table.name, self.name, vals)
             self.table.migration.operation.alter_column(
                 self.table.name, self.name,
                 schema=self.table.schema, **vals)
@@ -1251,6 +1302,7 @@ class Migration:
         self.metadata = registry.declarativebase.metadata
         self.ddl_compiler = self.conn.dialect.ddl_compiler(
             self.conn.dialect, None)
+        self.ignore_migration_for = registry.ignore_migration_for
 
         opts = {
             'include_schemas': True,
